+++ infra/postgres/migrations/0050_mcp_management_operation.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+DROP TABLE IF EXISTS mcp_management_operation;
+COMMIT;